Split zustand stores
