+++ packages/app/features/home/screen.tsx
@@ -73,7 +73,7 @@ export function HomeScreen() {
         pos="absolute"
         b={30}
         l={20}
-        icon={theme === 'dark' ? Sun : Moon}
+        icon={theme === 'dark' ? <Sun /> : <Moon />}
         onPress={toggleTheme}
         circular
       />
